Still more compiler warning fixes.
git-svn-id: svn+ssh://svn.gna.org/svn/warzone/branches/crossplatform@103 4a71c877-e1ca-e34f-864e-861f7616d084

+++ lib/sequence/adpcm.c
@@ -52,7 +52,7 @@ short adpcm_decode_sample(char code) {
 	}
 
 	/* Find new step_size index value */
-	step_idx += index_adjust[code];
+	step_idx += index_adjust[(int) code];
 
 	if (step_idx < 0) {
 		step_idx = 0;
